Seed only in `development` env

+++ db/seeds.rb
@@ -1,7 +1,9 @@
-AdminUser.destroy_all
+if Rails.env.development?
+  AdminUser.destroy_all
 
-AdminUser.create!(
-  email: 'admin@flipkart.com',
-  password: '123456',
-  password_confirmation: '123456'
-) if Rails.env.development?
+  AdminUser.create!(
+    email: 'admin@flipkart.com',
+    password: '123456',
+    password_confirmation: '123456'
+  )
+end
